third_party/WebKit/LayoutTests/fast/dom/HTMLTemplateElement/inertContents-expected.txt

The test asserts that elements within template contents are "inert", e.g. script does not run.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testVal is "script has not run"
PASS testVal is "script has run"
PASS successfullyParsed is true

TEST COMPLETE
